Add space find by name, fix 3.x issues and move Stream into separate file
